+++ src/WordPress/Shortcodes/ShowHomeClubShortcode.php
@@ -1,0 +1,34 @@
+<?php
+/**
+ * OpenTT - Table Tennis Management Plugin
+ * Copyright (C) 2026 Aleksa Dimitrijević
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ */
+
+namespace OpenTT\Unified\WordPress\Shortcodes;
+
+final class ShowHomeClubShortcode
+{
+    public static function render($atts, array $deps)
+    {
+        $call = static function ($name, ...$args) use ($deps) {
+            $name = (string) $name;
+            if (!isset($deps[$name]) || !is_callable($deps[$name])) {
+                return null;
+            }
+            return $deps[$name](...$args);
+        };
+
+        $ctx = $call('current_match_context');
+        if (!$ctx || empty($ctx['db_row'])) {
+            return '';
+        }
+
+        return (string) $call('shortcode_title_html', 'Domaćin')
+            . (string) $call('render_klub_card_html', intval($ctx['db_row']->home_club_post_id));
+    }
+}